refactor: offload logic to react hooks module
