+++ frontend/src/pages/agreements/approve/ApproveAgreement.jsx
@@ -1,7 +1,4 @@
-import React from "react";
-import { useNavigate, useParams, useSearchParams } from "react-router-dom";
 import App from "../../../App";
-import { useGetAgreementByIdQuery, useGetServicesComponentsListQuery } from "../../../api/opsAPI";
 import AgreementBLIAccordion from "../../../components/Agreements/AgreementBLIAccordion";
 import AgreementCANReviewAccordion from "../../../components/Agreements/AgreementCANReviewAccordion";
 import AgreementChangesAccordion from "../../../components/Agreements/AgreementChangesAccordion";
@@ -12,141 +9,43 @@ import ServicesComponentAccordion from "../../../components/ServicesComponents/S
 import TextArea from "../../../components/UI/Form/TextArea";
 import ConfirmationModal from "../../../components/UI/Modals/ConfirmationModal";
 import PageHeader from "../../../components/UI/PageHeader";
-import { BLI_STATUS, groupByServicesComponent } from "../../../helpers/budgetLines.helpers";
-import { getInReviewChangeRequests } from "../../../helpers/changeRequests.helpers";
+import { BLI_STATUS } from "../../../helpers/budgetLines.helpers";
 import { findDescription, findPeriodEnd, findPeriodStart } from "../../../helpers/servicesComponent.helpers";
-import { convertCodeForDisplay, renderField, toTitleCaseFromSlug } from "../../../helpers/utils";
-import useAlert from "../../../hooks/use-alert.hooks.js";
-import useToggle from "../../../hooks/useToggle";
-import useGetUserFullNameFromId from "../../../hooks/user.hooks";
-import { getTotalByCans } from "../review/ReviewAgreement.helpers";
+import { convertCodeForDisplay } from "../../../helpers/utils";
+import useApproveAgreement from "./ApproveAgreement.hooks";
 
 const ApproveAgreement = () => {
-    const { setAlert } = useAlert();
-    const urlPathParams = useParams();
-    const [notes, setNotes] = React.useState("");
-    const [confirmation, setConfirmation] = React.useState(false);
-    const [showModal, setShowModal] = React.useState(false);
-    const [modalProps, setModalProps] = React.useState({
-        heading: "",
-        actionButtonText: "",
-        secondaryButtonText: "",
-        handleConfirm: () => {}
-    });
-
-    const CHANGE_REQUEST_SLUG_TYPES = {
-        STATUS: "status-change",
-        BUDGET: "budget-change"
-    };
-    let submittersNotes = "This is a test note"; // TODO: replace with actual data
-    // @ts-ignore
-    const agreementId = +urlPathParams.id;
-    const [searchParams] = useSearchParams();
-    const navigate = useNavigate();
-    let changeRequestType = searchParams.get("type") ?? "";
-    let changeToStatus = searchParams.get("to")?.toUpperCase() ?? "";
-    const checkBoxText =
-        changeToStatus === BLI_STATUS.PLANNED
-            ? "I understand that approving these budget lines will subtract the amounts from the FY budget"
-            : "I understand that approving these budget lines will start the Procurement Process";
-    const approveModalHeading =
-        changeToStatus === BLI_STATUS.PLANNED
-            ? "Are you sure you want to approve these budget lines for Planned Status? This will subtract the amounts from the FY budget."
-            : "Are you sure you want to approve these budget lines for Executing Status? This will start the procurement process.";
-    const [afterApproval, setAfterApproval] = useToggle(true);
-
     const {
-        data: agreement,
-        error: errorAgreement,
-        isLoading: isLoadingAgreement
-    } = useGetAgreementByIdQuery(agreementId, {
-        refetchOnMountOrArgChange: true
-    });
+        agreement,
+        projectOfficerName,
+        servicesComponents,
+        groupedBudgetLinesByServicesComponent,
+        budgetLinesInReview,
+        changeRequestsInReview,
+        changeInCans,
+        notes,
+        setNotes,
+        confirmation,
+        setConfirmation,
+        showModal,
+        setShowModal,
+        modalProps,
+        checkBoxText,
+        handleCancel,
+        handleDecline,
+        handleApprove,
+        title,
+        changeRequestTitle,
+        afterApproval,
+        setAfterApproval,
+        submittersNotes,
+        changeToStatus,
+        statusForTitle
+    } = useApproveAgreement();
 
-    const projectOfficerName = useGetUserFullNameFromId(agreement?.project_officer_id);
-    const { data: servicesComponents } = useGetServicesComponentsListQuery(agreement?.id);
-
-    if (isLoadingAgreement) {
-        return <h1>Loading...</h1>;
+    if (!agreement) {
+        return <div>Loading...</div>;
     }
-    if (errorAgreement) {
-        return <h1>Oops, an error occurred</h1>;
-    }
-
-    // TODO: move this to a helper function
-    const groupedBudgetLinesByServicesComponent = groupByServicesComponent(agreement?.budget_line_items);
-    const budgetLinesInReview = agreement?.budget_line_items.filter((bli) => bli.in_review);
-    /**
-     *  @typedef {import('../../../components/ChangeRequests/ChangeRequestsList/ChangeRequests').ChangeRequest} ChangeRequest
-     *  @type {ChangeRequest[]}
-     */
-    const changeRequestsInReview = /** @type {ChangeRequest[]} */ (
-        getInReviewChangeRequests(agreement?.budget_line_items)
-    );
-    const changeInCans = getTotalByCans(budgetLinesInReview);
-
-    let statusForTitle = "";
-    if (changeRequestType === CHANGE_REQUEST_SLUG_TYPES.STATUS) {
-        const status = changeToStatus === "EXECUTING" ? BLI_STATUS.EXECUTING : BLI_STATUS.PLANNED;
-        statusForTitle = `- ${renderField(null, "status", status)}`;
-    }
-    const changeRequestTitle = toTitleCaseFromSlug(changeRequestType);
-    const title = `Approval for ${changeRequestTitle} ${statusForTitle}`;
-
-    const handleCancel = () => {
-        setShowModal(true);
-        setModalProps({
-            heading:
-                "Are you sure you want to cancel? This will exit the review process and you can come back to it later.",
-            actionButtonText: "Cancel",
-            secondaryButtonText: "Continue Reviewing",
-            handleConfirm: () => {
-                navigate("/agreements");
-            }
-        });
-    };
-
-    const declineChangeRequest = () => {
-        setAlert({
-            type: "success",
-            heading: "Not Yet Implemented",
-            message: "Not yet implemented"
-        });
-    };
-
-    const handleDecline = async () => {
-        setShowModal(true);
-        setModalProps({
-            heading: `Are you sure you want to decline these budget lines for ${changeToStatus} Status?`,
-            actionButtonText: "Decline",
-            secondaryButtonText: "Cancel",
-            handleConfirm: async () => {
-                await declineChangeRequest();
-                navigate("/agreements");
-            }
-        });
-    };
-
-    const approveChangeRequest = () => {
-        setAlert({
-            type: "success",
-            heading: "Not Yet Implemented",
-            message: "Not yet implemented"
-        });
-    };
-
-    const handleApprove = async () => {
-        setShowModal(true);
-        setModalProps({
-            heading: approveModalHeading,
-            actionButtonText: "Approve",
-            secondaryButtonText: "Cancel",
-            handleConfirm: async () => {
-                await approveChangeRequest();
-                await navigate("/agreements");
-            }
-        });
-    };
 
     return (
         <App breadCrumbName={`Approve BLI ${changeRequestTitle} ${statusForTitle}`}>
@@ -161,7 +60,7 @@ const ApproveAgreement = () => {
             )}
             <PageHeader
                 title={title}
-                subTitle={agreement.name}
+                subTitle={agreement?.name}
             />
             <ReviewChangeRequestAccordion
                 changeType={changeRequestTitle}
